+++ build.utils.js
@@ -81,6 +81,18 @@ clim.getTime = function(){
          now.getSeconds() + ']'));
 };
 
+clim.logWrite = function(level, prefixes, msg) {
+    // Default implementation writing to stderr
+    var line = clim.getTime() + " " + level;
+    if (prefixes.length > 0) line += " " + prefixes.join(" ");
+
+    line = colors.dim(line);
+    line += " " + msg;
+    process.stderr.write(line + "\n");
+
+    // or post it web service, save to database etc...
+  };
+
 /* Logic for inling styles adapted from rollup-plugin-angular CREDIT Felix Itzenplitz */
 
 function insertText(str, dir, preprocessor = res => res, processFilename = false) {
@@ -101,12 +113,19 @@ function insertText(str, dir, preprocessor = res => res, processFilename = false
 /* LOG Method used in the build tasks.
    Pretty prints LOG, magenta, green, blue, grey message */
 
-const log = function (action, noun, verb, next) {
-    let a = action ? colors.magenta(action) : '';
-    let n = noun ? colors.green(noun) : '';
-    let v = verb ? colors.cyan(verb) : '';
+const log = function (action, noun, next) {
+    let a = action ? colors.dim(colors.white(action)) : '';
+    let n = noun ? colors.dim(colors.green(noun)) : '';
     let x = next ? colors.dim(colors.white(next)) : '';
-    cons.log(a + ' ' + n + ' ' + v + ' ' + x );
+    cons.log(a + ' ' + n + ' ' + x );
+};
+
+const alert = function (noun, verb, action, next) {
+    let n = noun ?colors.magenta(noun) : '';
+    let v = verb ? colors.green(verb) : '';
+    let a = action ? colors.cyan(action) : '';
+    let x = next ? colors.dim(colors.white(next)) : '';
+    cons.log(n + ' ' + v + ' ' + a + ' ' + x );
 };
 
 /* WARN Method used in the build tasks.
@@ -152,6 +171,7 @@ const utils = {
     colors: colors,
     log : log,
     warn : warn,
+    alert: alert,
     moduleIdRegex: moduleIdRegex,
     directiveRegex: directiveRegex,
     componentRegex: componentRegex,