+++ main.prod.ts
@@ -1,7 +1,5 @@
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { platformBrowser } from '@angular/platform-browser';
 import { enableProdMode } from '@angular/core';
 import { AppModule } from './tmp/app/app.module';
 enableProdMode();
-platformBrowserDynamic().bootstrapModule(AppModule);
-
-
+platformBrowser().bootstrapModuleFactory(AppModule);